no need int64 to store number of columns

+++ src/data.table.h
@@ -153,8 +153,8 @@ int getDTthreads();
 void avoid_openmp_hang_within_fork();
 
 // roll.c
-void rollmeanVector(double x[], uint_fast64_t nx, double ans[], int k, double fill, bool exact, bool narm, bool hasna, bool nahasna);
-void rollmeanVectorAdaptive(double x[], uint_fast64_t nx, double ans[], int k[], double fill, bool exact, bool narm, bool hasna, bool nahasna);
+void rollmeanVector(double x[], int nx, double ans[], int k, double fill, bool exact, bool narm, bool hasna, bool nahasna);
+void rollmeanVectorAdaptive(double x[], int nx, double ans[], int k[], double fill, bool exact, bool narm, bool hasna, bool nahasna);
 
 // rollR.c
 SEXP rollmean(SEXP obj, SEXP k, SEXP fill, SEXP exact, SEXP align, SEXP narm, SEXP hasNA, SEXP adaptive);
